EO-29 On cart page, user can delete item / EO-32 Make a call to delete item endpoint

+++ app/src/main/java/cenergy/central/com/pwb_store/model/CartItem.kt
@@ -12,7 +12,7 @@ import io.realm.annotations.PrimaryKey
 open class CartItem(
         @PrimaryKey
         @SerializedName("item_id")
-        var id: Int? = 0,
+        var id: Long? = 0,
         var sku: String? = "",
         var qty: Int? = 0,
         var name: String? = "",
